Reorganize frontend assets by media type
