+++ frontend/public/favorites-movie.html
@@ -7,9 +7,9 @@
     <!-- CSS chung -->
     <link rel="stylesheet" href="/src/styles/base.css" />
     <!-- Layout card của phim -->
-    <link rel="stylesheet" href="/src/styles/components/movie-card.css" />
+    <link rel="stylesheet" href="/src/styles/components/movie/movie-card.css" />
     <link rel="icon" type="image/png" href="/default/favicon.png" />
-    <script type="module" src="/src/pages/favorites-movie.js"></script>
+    <script type="module" src="/src/pages/movie/favorites.js"></script>
   </head>
   <body>
     <!-- HEADER -->
