+++ frontend/public/favorites.html
@@ -5,11 +5,11 @@
     <title>❤️ Yêu thích</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/src/styles/base.css" />
-    <link rel="stylesheet" href="/src/styles/pages/index.css" />
+    <link rel="stylesheet" href="/src/styles/pages/manga/index.css" />
     <link rel="stylesheet" href="/src/styles/components/folder-card.css" />
      <link rel="stylesheet" href="/src/styles/components/folder-slider.css" />
     <link rel="icon" type="image/png" href="/default/favicon.png" />
-    <script type="module" src="/src/pages/favorites.js"></script>
+    <script type="module" src="/src/pages/manga/favorites.js"></script>
   </head>
   <body>
     <header id="site-header" class="site-header">
