+++ frontend/public/index.html
@@ -6,11 +6,11 @@
     <link rel="icon" type="image/png" href="/default/favicon.png" />
     <title>📚 MyLocalManga</title>
     <link rel="stylesheet" href="/src/styles/base.css" />
-    <link rel="stylesheet" href="/src/styles/pages/index.css" />
+    <link rel="stylesheet" href="/src/styles/pages/manga/index.css" />
     <link rel="stylesheet" href="/src/styles/dark/home-dark.css" />
     <link rel="stylesheet" href="/src/styles/components/folder-slider.css" />
     <link rel="stylesheet" href="/src/styles/components/folder-card.css" />
-    <script type="module" src="/src/pages/index.js"></script>
+    <script type="module" src="/src/pages/manga/index.js"></script>
   </head>
   <body>
     <!-- ✅ HEADER GỐC -->
