+++ frontend/public/movie-index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Danh sách phim</title>
     <link rel="stylesheet" href="/src/styles/base.css" />
-    <link rel="stylesheet" href="/src/styles/components/movie-card.css" />
+    <link rel="stylesheet" href="/src/styles/components/movie/movie-card.css" />
     <link rel="stylesheet" href="/src/styles/components/folder-slider.css" />
   </head>
 
@@ -80,6 +80,6 @@
     <div id="loading-overlay" class="hidden">
       <div class="loader-text">⏳ Đang tải...</div>
     </div>
-    <script type="module" src="/src/pages/movie-index.js"></script>
+    <script type="module" src="/src/pages/movie/index.js"></script>
   </body>
 </html>
