+++ frontend/public/movie-player.html
@@ -5,9 +5,9 @@
     <title>Xem phim</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/src/styles/base.css" />
-    <link rel="stylesheet" href="/src/styles/components/movie-card.css" />
+    <link rel="stylesheet" href="/src/styles/components/movie/movie-card.css" />
     <link rel="stylesheet" href="/src/styles/components/folder-slider.css" />
-    <link rel="stylesheet" href="/src/styles/pages/player.css" />
+    <link rel="stylesheet" href="/src/styles/pages/movie/player.css" />
 
     <link rel="icon" type="image/png" href="/default/favicon.png" />
   </head>
@@ -75,7 +75,7 @@
     </div>
 
     <div id="loading-overlay" class="hidden">⏳ Đang tải...</div>
-    <script type="module" src="/src/pages/movie-player.js"></script>
+    <script type="module" src="/src/pages/movie/player.js"></script>
     <script type="module" src="/src/core/ui.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/hammerjs@2.0.8/hammer.min.js"></script>
