+++ frontend/public/music-index.html
@@ -6,8 +6,8 @@
     <title>Danh sách nhạc</title>
     <link rel="stylesheet" href="/src/styles/base.css" />
     <link rel="stylesheet" href="/src/styles/components/folder-slider.css" />
-    <link rel="stylesheet" href="/src/styles/components/music-card.css" />
-    <link rel="stylesheet" href="/src/styles/components/play-list-popup.css" />
+    <link rel="stylesheet" href="/src/styles/components/music/music-card.css" />
+    <link rel="stylesheet" href="/src/styles/components/music/play-list-popup.css" />
   </head>
   <body>
     <header id="site-header" class="site-header">
@@ -77,6 +77,6 @@
 <div id="loading-overlay" class="hidden">
   <div class="loader-text">⏳ Đang tải...</div>
 </div>
-    <script type="module" src="/src/pages/music-index.js"></script>
+    <script type="module" src="/src/pages/music/index.js"></script>
   </body>
 </html>
