+++ frontend/public/playlist-music.html
@@ -5,7 +5,7 @@
     <title>Playlist</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/src/styles/base.css" />
-    <link rel="stylesheet" href="/src/styles/components/movie-card.css" />
+    <link rel="stylesheet" href="/src/styles/components/movie/movie-card.css" />
   </head>
   <body>
     <header class="site-header">
@@ -25,6 +25,6 @@
       <div id="playlist-container" class="grid"></div>
     </div>
 
-    <script type="module" src="/src/pages/playlist-music.js"></script>
+    <script type="module" src="/src/pages/music/playlist.js"></script>
   </body>
 </html>
