+++ frontend/public/reader.html
@@ -6,7 +6,7 @@
     <link rel="icon" type="image/png" href="/default/favicon.png" />
     <title>Đọc truyện - MyLocalManga</title>
     <link rel="stylesheet" href="/src/styles/base.css" />
-    <link rel="stylesheet" href="/src/styles/pages/reader.css" />
+    <link rel="stylesheet" href="/src/styles/pages/manga/reader.css" />
     <link rel="stylesheet" href="/src/styles/dark/reader-dark.css" />
     <!-- Thư viện SwiperJS + PinchZoom -->
 
@@ -70,6 +70,6 @@
       <div class="loader-text">⏳ Đang tải...</div>
     </div>
     <!-- ✅ Load JS đúng cách -->
-    <script type="module" src="/src/pages/reader.js"></script>
+    <script type="module" src="/src/pages/manga/reader.js"></script>
   </body>
 </html>
