+++ frontend/src/components/movie/movieCard.js
