+++ frontend/src/pages/manga/favorites.js
