+++ frontend/src/pages/manga/index.js
