+++ frontend/src/pages/manga/reader.js
