+++ frontend/src/pages/movie/player.js
