+++ frontend/src/styles/components/movie/movie-card.css
