+++ frontend/src/styles/components/music/music-card.css
