+++ frontend/src/styles/components/music/play-list-popup.css
