+++ frontend/src/styles/pages/manga/index.css
