+++ frontend/src/styles/pages/manga/reader.css
