+++ frontend/src/styles/pages/movie/player.css
