+++ frontend/src/styles/pages/music/player.css
